Boilerplate setup for truffle and GH actions CI
This commit initializes the basic structure for the code, by making use
of truffle.
TODO: the truffle config file requires to specify the network we host
the local blockchain on (e.g. localhost and a port). This needs to be
done once it's up, after which we can also add actual tests and our
contracts.